Cleanup and small refactor
